Fix bug grecaptcha is undefined
Downloading a vip content require grecaptcha but after getting script the object return null most of the times. However, while checking window.grecaptcha object, it's loaded. This might be due to the permission grant so I've changed it to auto instead of content.
Besides, right click does not work, and it's not always have "Đọc Từ Đầu" string, I've made a minor update on this too.